ISPN-9100 Move doclets to dedicated jar
- only build the doclets when building the distribution
- Fixes the build with JDK10+
