+++ javadoc/doclets/src/main/java/org/infinispan/doclets/jmx/JmxComponent.java
@@ -1,4 +1,4 @@
-package org.infinispan.tools.doclet.jmx;
+package org.infinispan.doclets.jmx;
 
 /**
  * A JMX Component
