+++ javadoc/javadoc-all/pom.xml
@@ -190,10 +190,10 @@
                      <goal>jar</goal>
                   </goals>
                   <configuration>
-                     <doclet>org.infinispan.tools.doclet.DocletMultiplexer</doclet>
+                     <doclet>org.infinispan.doclets.DocletMultiplexer</doclet>
                      <docletArtifact>
                         <groupId>org.infinispan</groupId>
-                        <artifactId>infinispan-tools</artifactId>
+                        <artifactId>infinispan-doclets</artifactId>
                         <version>${project.version}</version>
                      </docletArtifact>
                      <doctitle>${infinispan.brand.name} ${infinispan.base.version} API</doctitle>
@@ -321,7 +321,7 @@
             <dependencies>
                <dependency>
                   <groupId>org.infinispan</groupId>
-                  <artifactId>infinispan-tools</artifactId>
+                  <artifactId>infinispan-doclets</artifactId>
                   <version>${project.version}</version>
                </dependency>
             </dependencies>
