+++ pom.xml
@@ -202,6 +202,7 @@
       <module.skipTestComponentMetaDataProcessing>true</module.skipTestComponentMetaDataProcessing>
       <module.skipMavenRemoteResource>false</module.skipMavenRemoteResource>
       <module.skipCopyDependencies>true</module.skipCopyDependencies>
+      <module.skipSupportModuleDeployment>true</module.skipSupportModuleDeployment>
       <packaging>jar</packaging>
       <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
       <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -2035,6 +2036,7 @@
       <profile>
          <id>distribution</id>
          <modules>
+            <module>javadoc/doclets</module>
             <module>javadoc/javadoc-embedded</module>
             <module>javadoc/javadoc-remote</module>
             <module>javadoc/javadoc-all</module>
