+++ server/integration/versions/pom.xml
@@ -479,7 +479,6 @@
          <plugin>
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-enforcer-plugin</artifactId>
-            <version>1.3.1</version>
             <executions>
                <execution>
                   <id>enforce-java</id>
@@ -492,7 +491,7 @@
                            <version>[${version.java},)</version>
                         </requireJavaVersion>
                         <requireMavenVersion>
-                           <version>[3.0.5,)</version>
+                           <version>[3.2.1,)</version>
                         </requireMavenVersion>
                      </rules>
                   </configuration>
